corrected function casing style according to golang docs
An identifier may be exported to permit access to it from another package. An identifier is exported if both:

the first character of the identifier's name is a Unicode uppercase letter (Unicode character category Lu); and
the identifier is declared in the package block or it is a field name or method name.

Explained here: https://stackoverflow.com/questions/38616687/which-way-to-name-a-function-in-go-camelcase-or-semi-camelcase